+++ src/Gnuplot.jl
@@ -14,7 +14,7 @@ export session_names, palette_names, linetypes, palette_levels, palette,
 
 Return the **Gnuplot.jl** package version.
 """
-version() = v"1.6.3"
+version() = v"1.6.4"
 
 
 # ---------------------------------------------------------------------